Refactor authentication related code.
Refactor was done as a backward compatible change with
TODO notes for completion.

See ARCH-244 for details on completing this work.

ARCH-244
