+++ edx_rest_framework_extensions/__init__.py
@@ -1,3 +1,3 @@
 """ edx Django REST Framework extensions. """
 
-__version__ = '1.7.0'  # pragma: no cover
+__version__ = '1.8.0'  # pragma: no cover
