+++ edx_rest_framework_extensions/auth/README.rst
@@ -1,0 +1,5 @@
+This directory contains authentication related extensions.
+
+See `ARCH-244`_ for details of completing the refactor into this directory.
+
+.. _ARCH-244: https://openedx.atlassian.net/browse/ARCH-244
