--- edx_rest_framework_extensions/authn/README.rst
@@ -1,8 +1,0 @@
-This directory will contain all authentication related extensions.
-
-It contains the following subdirectory:
-- jwt
-
-**Note:** The old JWT and Bearer authentication code has not yet been
-refactored into this directory. That update is intended to come
-imminently.
